Implementados y arreglados CAs 3 y 5 (pedidos que se borran al confirmar por falta de stock y pedidos no confirmados que se borran si alguien más confirma y deja sin stock). stockDeItemsSinConfirmar implementado en fachada y subsistema servicio, es llamado desde el controlador de dispositivo cuando se reciben eventos de cambio de estado de pedidos.

+++ src/Modelo/EstadosDePedido/Pedido.java
@@ -62,10 +62,12 @@ public class Pedido {
     }
     
     protected void hacerEliminarPedido(){
-        this.item.devolverStockDeInsumos();
+        if (estado.getTipo() != EstadoPedido.TipoDeEstado.NO_CONFIRMADO) {
+            this.item.devolverStockDeInsumos();            
+            this.getItem().getUnidadProcesadora().eliminarPedido(this);
+        }
         servicio.removerPedido(this);
         this.servicio = null;
-        this.getItem().getUnidadProcesadora().eliminarPedido(this);
     }
     
     protected void hacerConfirmarPedido() throws PolloException {
